@@ -1,5 +1,5 @@
 module github.com/ramonmoraes/timer
 
-go 1.13
+go 1.16
 
 require github.com/faiface/beep v1.0.2
